Got form validation semi-working
