+++ src/components/CreateListing/CreateListing.js
@@ -16,6 +16,7 @@ import Uploader from '../Uploader/Uploader';
 import { createListing, getUserId } from '../../actions';
 import ImagePreview from './ImagePreview';
 import cameraBanner from '../../Images/banner.png';
+import Form from './Form';
 
 const CreateListing = () => {
   const dispatch = useDispatch();
@@ -30,45 +31,75 @@ const CreateListing = () => {
   );
 
   // react state
-  const [name, setName] = useState('');
+  const [name] = useState('');
   const [picture, setPicture] = useState('');
-  const [price, setPrice] = useState('');
-  const [city, setCity] = useState('');
-  const [state, setLocation] = useState('');
-  const [zipcode, setZipcode] = useState('');
-  const [category, setCategory] = useState('');
-  const [subcategory, setSubCat] = useState('');
-  const [description, setDescription] = useState('');
-  const [paymentType, setPayment] = useState('');
+  const [price] = useState('');
+  const [city] = useState('');
+  const [state] = useState('');
+  const [zipcode] = useState('');
+  const [category] = useState('');
+  const [subcategory] = useState('');
+  const [description] = useState('');
+  const [paymentType] = useState('');
   const [count, setCount] = useState(0);
   const [available] = useState(true);
   const [average_rating] = useState(0);
-  const [condition, setCondition] = useState('');
+  const [condition] = useState('');
 
   // this is called to get the user id.
   useEffect(() => {
     dispatch(getUserId(userToken));
   }, []);
 
-  const listing = {
+  const handleSubmit = (id, list) => {
+    dispatch(createListing(id, list));
+  };
+
+  const item = {
     users_ownerId: userId,
+    name,
+    picture,
+    price,
+    city,
+    state,
+    zipcode,
+    category,
+    sub_category: subcategory,
+    description,
+    payment_type: paymentType,
+    available,
+    average_rating,
+    condition,
+  };
+
+  const listing = {
+    userId,
     name,
     price,
     picture,
     category,
     description,
     available,
-    payment_type: paymentType,
+    paymentType,
     average_rating,
     condition,
-    sub_category: subcategory,
+    subcategory,
     city,
     state,
     zipcode,
+    setPicture,
+    setCount,
+    count,
+    handleSubmit,
+    isSubmitting,
   };
 
-  const listingArr = Object.values(listing);
-  const listingArrFilter = listingArr.filter(item => item !== '');
+  // const listingArr = Object.values(listing);
+  // const listingArrFilter = listingArr.filter(item => item !== '');
+
+  // console.log(listingArrFilter);
+  // console.log(listing);
+  console.log(item);
 
   return (
     <div className="create-listing">
@@ -76,189 +107,7 @@ const CreateListing = () => {
         <img src={cameraBanner} alt="" />
       </div>
       <div className="form-body">
-        <form>
-          {/* conditional render for image preview, will change this later on.  */}
-          <br />
-          <div className="left-side">
-            <div className="image-items">
-              {picture ? (
-                <div className="image-preview">
-                  <ImagePreview image={picture} count={count} />
-                </div>
-              ) : (
-                <div className="image-holder">
-                  <p>
-                    Click Below <br />
-                    to <br />
-                    Add Images
-                  </p>
-                </div>
-              )}
-              <br />
-              {/* The uploadcare uploader */}
-              <Uploader
-                className="upload"
-                id="picture"
-                onUploadComplete={info => {
-                  setCount(info.count);
-                  setPicture(info.uuid);
-                }}
-              />
-            </div>
-            <div className="description-div">
-              Description{' '}
-              <textarea
-                name="description"
-                value={description}
-                type="text"
-                onChange={e => setDescription(e.target.value)}
-                className="description"
-                required
-              />
-            </div>
-          </div>
-          <div className="right-side">
-            Product{' '}
-            <input
-              name="name"
-              value={name}
-              type="text"
-              onChange={e => setName(e.target.value)}
-              className="long-input"
-            />
-            Price{' '}
-            <input
-              name="price"
-              value={price}
-              type="number"
-              onChange={e => setPrice(e.target.value)}
-              className="long-input"
-            />
-            <div className="middle-row">
-              <div className="city-input-field">
-                City{' '}
-                <input
-                  name="city"
-                  value={city}
-                  type="text"
-                  onChange={e => setCity(e.target.value)}
-                  className="medium-input"
-                />
-              </div>
-              <div className="middle-row-field">
-                State
-                {/* put this into its own component */}
-                <StateDropDown
-                  handleChange={e => setLocation(e.target.value)}
-                />
-              </div>
-              <div className="middle-row-field">
-                Zipcode{' '}
-                <input
-                  name="zipcode"
-                  value={zipcode}
-                  type="number"
-                  onChange={e => setZipcode(e.target.value)}
-                  className="small-input"
-                />
-              </div>
-            </div>
-            <div>
-              Category{' '}
-              <select
-                name="category"
-                value={category}
-                type="text"
-                onChange={e => setCategory(e.target.value)}
-                className="long-input"
-              >
-                <option value="" disabled>
-                  Choose Category
-                </option>
-                <option>Mounts</option>
-                <option>Cameras</option>
-                <option>Lenses</option>
-                <option>Lighting</option>
-                <option>Support Equipment</option>
-                <option>Accessories</option>
-              </select>
-            </div>
-            {/* very long file so I refactored it as its own component in ./SubCategory.js */}
-            <SubCategory
-              category={category}
-              handleChange={e => setSubCat(e.target.value)}
-            />
-            <div>
-              Condition <br />
-              <select
-                name="condition"
-                value={condition}
-                type="text"
-                onChange={e => setCondition(e.target.value)}
-                className="long-input"
-              >
-                <option value="" disabled>
-                  Choose Condition
-                </option>
-                <option>Like New</option>
-                <option>Used (normal wear)</option>
-                <option>Other (see description)</option>
-              </select>
-            </div>
-          </div>
-        </form>
-        <div className="bottom-row">
-          <div className="payment">
-            Payment Preference
-            <div className="payment-options">
-              <div className="option">
-                <input
-                  name="paymentType"
-                  value="cash"
-                  type="radio"
-                  onChange={e => setPayment(e.target.value)}
-                  checked={paymentType === 'cash'}
-                />{' '}
-                Cash
-              </div>
-              <div className="option">
-                <input
-                  name="paymentType"
-                  value="card"
-                  type="radio"
-                  onChange={e => setPayment(e.target.value)}
-                  checked={paymentType === 'card'}
-                />{' '}
-                Card
-              </div>
-              <div className="option">
-                <input
-                  name="paymentType"
-                  value="both"
-                  type="radio"
-                  onChange={e => setPayment(e.target.value)}
-                  checked={paymentType === 'both'}
-                />{' '}
-                Both
-              </div>
-            </div>
-          </div>
-          <div className="lower-buttons">
-            <button className="cancel">Cancel</button>
-            <button
-              onClick={e => {
-                e.preventDefault();
-                dispatch(createListing(userId, listing));
-              }}
-              className={`${
-                listingArrFilter.length === 14 ? 'list' : 'list disabled'
-              }`}
-              disabled={listingArrFilter.length < 14 || isSubmitting}
-            >
-              List Item
-            </button>
-          </div>
-        </div>
+        <Form listing={listing} item={item} />
       </div>
     </div>
   );
