+++ src/components/CreateListing/_createListing.scss
@@ -204,8 +204,8 @@
 
 .uploadcare--widget__file-name {
   color: #fff;
-  margin-left: -30%;
   font-size: 40px;
+  margin-left: 18%;
   background-color: #000000;
   border-radius: 3px;
   text-decoration: none;
